pyproject.toml
[project]
name = "weatherxm-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-mcp>=0.3.3",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "PyYAML>=6.0",
]
